Added tech/frameworks to software section. Fixed nav links.
Went with a different approach instead of using native
'smooth-behavior: smooth' provided by css. Using links
was causing the back button to update with the id of
the div container. Therefore, one must go back several
times before leaving the page which is undesirable.
With this approach the url is never updated and so
the back button is never updated.